github_actions/requirements.txt
Django==4.1.5
django-cors-headers==3.10.1
djangorestframework==3.13.1
pymysql
mysqlclient